Seperate the patches into files for each class that we're patching
