+++ ModTek/Patches/MetadataDatabase.cs
@@ -1,0 +1,21 @@
+using BattleTech.Data;
+using Harmony;
+
+namespace ModTek
+{
+    /// <summary>
+    /// Patch the MDDB path to direct to the one in the .modtek path
+    /// </summary>
+    [HarmonyPatch(typeof(MetadataDatabase))]
+    [HarmonyPatch("MDD_DB_PATH", MethodType.Getter)]
+    public static class MetadataDatabase_MDD_DB_PATH_Patch
+    {
+        public static void Postfix(ref string __result)
+        {
+            if (string.IsNullOrEmpty(ModTek.ModMDDBPath))
+                return;
+
+            __result = ModTek.ModMDDBPath;
+        }
+    }
+}
